Move KryptDataBase to database module
